Switch from 'toml' package to tomli/tomllib
Python 3.11 comes with built-in support for TOML. It's an almost direct
import of the tomli package. As the 'toml' library is unmaintained, and
tomli has the same interface as what's in the standard library, this
switches the TOML support to a supported library with a minimum of fuss.

Tomllib requires files to be opened in 'rb' mode, so the relevant tests
had to be updated accordingly.