Install and use toml conditionally on Python2
tomli is not available for Python2, so to keep the Python2
compatibility, we need to install and use toml instead.

Use conditional import on ImportError to avoid introducing tox as new
dependency.

Unfortunately toml's API is slightly different than tomli's and
tomllib's, so we need additional implementation of the toml loading in
case tomllib didn't work.