Install and use toml conditionally on Python2 (#88)
tomli is not available for Python2, so to keep the Python2
compatibility, we need to install and use toml instead.

Use conditional import on ImportError to avoid introducing tox as new
dependency.

As there is difference in the naming of the toml and tomli error
classes, we use generic Exception. More over, we actually want to
fallback to empty config nevertheless what went wrong.